# speedup for make_gaussian_blobs