--- a/examples/simulated/plot_classifier_comparison.py
+++ b/examples/simulated/plot_classifier_comparison.py
@@ -224,11 +224,11 @@ datasets = [
     ),
     make_gaussian_blobs(
         2*n_matrices, n_channels, random_state=rs, class_sep=1., class_disp=.2,
-        n_jobs=-1
+        n_jobs=4
     ),
     make_gaussian_blobs(
         2*n_matrices, n_channels, random_state=rs, class_sep=.5, class_disp=.5,
-        mat_std=.1, n_jobs=-1
+        mat_std=.1, n_jobs=4
     )
 ]
 n_datasets = len(datasets)