--- a/tests/test_sampling.py
+++ b/tests/test_sampling.py
@@ -7,11 +7,14 @@ from pyriemann.utils.distance import distance_riemann
 from pyriemann.utils.test import is_sym_pos_def as is_spd
 
 
-def test_sample_gaussian_spd():
+@pytest.mark.parametrize("n_jobs", [1, -1])
+def test_sample_gaussian_spd(n_jobs):
     """Test Riemannian Gaussian sampling."""
-    n_matrices, n_dim, sigma = 50, 16, 2.
+    n_matrices, n_dim, sigma = 10, 8, 1.
     mean = np.eye(n_dim)
-    X = sample_gaussian_spd(n_matrices, mean, sigma, random_state=42)
+    X = sample_gaussian_spd(
+        n_matrices, mean, sigma, random_state=42, n_jobs=n_jobs
+        )
     assert X.shape == (n_matrices, n_dim, n_dim)  # X shape mismatch
     assert is_spd(X)  # X is an array of SPD matrices
 
@@ -24,12 +27,17 @@ def test_generate_random_spd_matrix():
     assert is_spd(X)  # X is a SPD matrix
 
 
-def test_sigma_gaussian_spd():
+@pytest.mark.parametrize("n_jobs", [1, -1])
+def test_sigma_gaussian_spd(n_jobs):
     """Test sigma parameter from Riemannian Gaussian sampling."""
-    n_matrices, n_dim, sig_1, sig_2 = 50, 8, 1., 2.
+    n_matrices, n_dim, sig_1, sig_2 = 10, 8, 1., 2.
     mean = np.eye(n_dim)
-    X1 = sample_gaussian_spd(n_matrices, mean, sig_1, random_state=42)
-    X2 = sample_gaussian_spd(n_matrices, mean, sig_2, random_state=66)
+    X1 = sample_gaussian_spd(
+        n_matrices, mean, sig_1, random_state=42, n_jobs=n_jobs
+        )
+    X2 = sample_gaussian_spd(
+        n_matrices, mean, sig_2, random_state=66, n_jobs=n_jobs
+        )
     avg_d1 = np.mean([distance_riemann(X1_i, mean) for X1_i in X1])
     avg_d2 = np.mean([distance_riemann(X2_i, mean) for X2_i in X2])
     assert avg_d1 < avg_d2