`make_gaussian_blobs` is a really nice addition to pyRiemann, but it is computationally expensive. The complexity depends mainly on the number of matrices to generate. The sampling code is a loop that generate one matrix at each iteration. As the samples are indenpendent, we could easily parallelize the process to reduce computation time. This is the object of this PR.

NB: this could reduce a lot the doc building time:
```
computation time summary:
    - ../examples/simulated/plot_classifier_comparison.py:                235.60 sec   0.0 MB
    - ../examples/stats/plot_oneWay_Manova.py:                             67.49 sec   0.0 MB
    - ../examples/artifacts/plot_detect_riemannian_potato_field_EEG.py:    49.95 sec   0.0 MB
    - ../examples/artifacts/plot_detect_riemannian_potato_EEG.py:          37.31 sec   0.0 MB
    - ../examples/simulated/plot_toy_classification.py:                    15.94 sec   0.0 MB
    - ../examples/ERP/plot_classify_MEG_mdm.py:                            11.12 sec   0.0 MB
    - ../examples/ERP/plot_embedding_MEG.py:                               11.03 sec   0.0 MB
...
```