@agramfort 2022-06-21
@sylvchev can you say how much n_jobs helps by giving running time you obtain? 🙏