@sylvchev 2022-06-23
> @sylvchev can you say how much n_jobs helps by giving running time you obtain? 🙏

for plot_toy_classification.py, with n_jobs=1:
```13 s ± 438 ms per loop (mean ± std. dev. of 3 runs, 1 loop each)```
and with n_jobs=8:
```5.88 s ± 940 ms per loop (mean ± std. dev. of 3 runs, 1 loop each)```