@agramfort 2022-07-07
@plcrodrigues merge if happy :)